Update newPygrep.py
Changed regex class constructor's openFile call to reference the instantiated regex object's openType, instead of the FileHandler class' openType, which wasn't set.

Changed readFile method in FileHandler, to write the contents of the file into the self.fileContents variable, so the whole file is in a string. 

We changed configList's re.findall call's second argument to be brx.fileContents, and ahead of the configList entry we perform the brx.readFile() method to populate the brx.fileContents variable.